requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
langchain-core==0.1.31
pydantic==2.6.3
python-dotenv==1.0.1
streamlit==1.32.0
openai==1.12.0 